implement unhappy scenarios

+++ src/main/java/com/github/maleksandrowicz93/cqrsdemo/student/DeleteStudentCommandHandler.java
@@ -1,5 +1,6 @@
 package com.github.maleksandrowicz93.cqrsdemo.student;
 
+import com.github.maleksandrowicz93.cqrsdemo.student.exception.StudentNotFoundException;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
@@ -7,7 +8,10 @@ class DeleteStudentCommandHandler {
 
     private final StudentRepository studentRepository;
 
-    boolean handle(long studentId) {
+    boolean handle(long studentId) throws StudentNotFoundException {
+        if (!studentRepository.existsById(studentId)) {
+            throw new StudentNotFoundException();
+        }
         studentRepository.deleteById(studentId);
         return true;
     }
